+++ src/rox_icu/cli.py
@@ -29,16 +29,21 @@ def info() -> None:
 @click.option("--node-id", default=1, help="device ID")
 @click.option("--sim-inputs", is_flag=True, help="simulate inputs")
 @click.option("--mqtt-broker", help="MQTT broker address", default=None)
+@click.option("--can-bus", help="CAN bus name", default="env")
 def mock(
     node_id: int,
     sim_inputs: bool,
     mqtt_broker: str | None,
+    can_bus: str,
 ):
     """Mock ICU device on CAN bus"""
     from .mock import ICUMock
 
     def main() -> None:
-        dev = ICUMock(node_id, simulate_inputs=sim_inputs, mqtt_broker=mqtt_broker)
+        bus = get_can_bus(can_bus)
+        dev = ICUMock(
+            node_id, simulate_inputs=sim_inputs, mqtt_broker=mqtt_broker, can_bus=bus
+        )
         dev.start()
 
     run_main(main)
@@ -79,7 +84,7 @@ def output(node_id: int, hex_input: str) -> None:
     print(f"Setting ICU_{node_id} output state to {state:08b}")
 
     with get_can_bus() as bus:
-        arb_id, data = canp.encode_message(canp.IoStateMessage(1, state), node_id)
+        arb_id, data = canp.encode_message(canp.SetIoStateMessage(state), node_id)
         msg = can.Message(arbitration_id=arb_id, data=data, is_extended_id=False)
         bus.send(msg)